bin: move TlsCertConfig out of dnssec module
It is unrelated to DNSSEC functionality.
