+++ bin/src/dnssec.rs
@@ -7,11 +7,9 @@
 
 //! Configuration types for all security options in hickory-dns
 
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
-#[cfg(feature = "dns-over-rustls")]
-use rustls::pki_types::{CertificateDer, PrivateKeyDer};
-#[cfg(all(feature = "dnssec-ring", not(feature = "dns-over-rustls")))]
+#[cfg(feature = "dnssec-ring")]
 use rustls_pki_types::PrivateKeyDer;
 use serde::Deserialize;
 #[cfg(feature = "dnssec-ring")]
@@ -221,63 +219,6 @@ impl KeyConfig {
                 .try_into()
                 .map_err(|e| format!("error converting time to std::Duration: {e}"))?,
         ))
-    }
-}
-
-/// Configuration for a TLS certificate
-#[derive(Deserialize, PartialEq, Eq, Debug)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TlsCertConfig {
-    pub path: PathBuf,
-    pub endpoint_name: Option<String>,
-    pub private_key: PathBuf,
-}
-
-impl TlsCertConfig {
-    /// Load a Certificate from the path (with rustls)
-    #[cfg(feature = "dns-over-rustls")]
-    pub fn load(
-        &self,
-        zone_dir: &Path,
-    ) -> Result<(Vec<CertificateDer<'static>>, PrivateKeyDer<'static>), String> {
-        use std::ffi::OsStr;
-
-        use tracing::info;
-
-        use hickory_proto::rustls::tls_server::{read_cert, read_key, read_key_from_der};
-
-        if self.path.extension().and_then(OsStr::to_str) != Some("pem") {
-            return Err(format!(
-                "unsupported certificate file format (expected `.pem` extension): {}",
-                self.path.display()
-            ));
-        }
-
-        let cert_path = zone_dir.join(&self.path);
-        info!(
-            "loading TLS PEM certificate chain from: {}",
-            cert_path.display()
-        );
-        let cert_chain = read_cert(&cert_path).map_err(|e| format!("error reading cert: {e}"))?;
-
-        let key_extension = self.private_key.extension();
-        let key = if key_extension.is_some_and(|ext| ext == "pem") {
-            let key_path = zone_dir.join(&self.private_key);
-            info!("loading TLS PKCS8 key from PEM: {}", key_path.display());
-            read_key(&key_path)?
-        } else if key_extension.is_some_and(|ext| ext == "der" || ext == "key") {
-            let key_path = zone_dir.join(&self.private_key);
-            info!("loading TLS PKCS8 key from DER: {}", key_path.display());
-            read_key_from_der(&key_path)?
-        } else {
-            return Err(format!(
-                "unsupported private key file format (expected `.pem` or `.der` extension): {}",
-                self.private_key.display()
-            ));
-        };
-
-        Ok((cert_chain, key))
     }
 }
